Minor fixes and rearrangements
Moved the ImagickCaptcha to the new Form factory.
Added current state of Form factory.
Removed Controller\Action's autoloader code, instead it now registers
the Application\Autoloader for the current module, as it removes the
latency of re-building the autoloader.
Changed the User Agent of Mail\Send
Added new exception, Exception\Length which will be thrown when a
specific length is expected.